refactor: rename project from resma-cli to resma
