+++ resma/images.py
